Fix BigDecimal ops, nullary builtins, and SIGN
Several fixes to codegen, typechecking and stdlib to prevent verifier errors and align runtime types:

- Match range comparisons for DECIMAL/MONEY now use BigDecimal.compareTo (not String.compareTo), preventing VerifyError when matching on BigDecimal/Money.
- Add emitDecimalPow helper and route the `**` operator and POWER builtin through it; exponent is narrowed with intValueExact() so fractional/oversized exponents throw rather than silently floor.
- POWER builtin lowered to use BigDecimal helper (matching the typechecker) instead of the integer overload.
- Recognise zero-argument builtins (TODAY, NOW, UUID-GENERATE, UUID-NIL) as NULLARY_BUILTINS in the typechecker and lower bare references to those names to builtin calls in codegen, allowing field initializers like `d : DATE = TODAY`.
- SIGN now returns Long to match Kobol INTEGER (JVM long); updated KobolMath, codegen callsites and tests accordingly.
- Update TypeChecker hint strings and add unit/integration tests covering POWER, SIGN, MONEY match ranges, and bare nullary builtin initializers.
- Bump kobolVersion to 0.1.4.

These changes fix class-load VerifyErrors caused by descriptor/type mismatches and ensure decimal exactness and proper typing across builtin lowering and code generation.

+++ stdlib/src/main/kotlin/dev/kobol/stdlib/KobolMath.kt
@@ -91,8 +91,11 @@ object KobolMath {
 
     // ── Sign ──────────────────────────────────────────────────────────────
 
-    @JvmStatic fun sign(n: Long): Int       = n.compareTo(0L)
-    @JvmStatic fun sign(n: BigDecimal): Int = n.signum()
+    // Kobol INTEGER is a JVM `long`, so SIGN returns Long (not Int) — the result is
+    // stored directly into an INTEGER slot with no widening (#v8: the Int overload
+    // produced an `int` that the verifier rejected in a `long_2nd` slot).
+    @JvmStatic fun sign(n: Long): Long       = n.compareTo(0L).toLong()
+    @JvmStatic fun sign(n: BigDecimal): Long = n.signum().toLong()
 
     // ── Rounding — Long-scale overloads (Kobol INTEGER is JVM long) ──────
